We now hava an application which manages its dependencies on its own.
