+++ gormExample.go
@@ -7,7 +7,7 @@ import (
 )
 
 func gormExample() {
-	db, err := gorm.Open(sqlite.Open("test.db"), &gorm.Config{})
+	db, err := gorm.Open(sqlite.Open("test.DB"), &gorm.Config{})
 	if err != nil {
 		panic("failed to connect database")
 	}
